% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anidata.R
\docType{data}
\name{anidata}
\alias{anidata}
\title{An example data to demonstrate AniPlot() function}
\format{A data frame with 72 rows and 4 variables:
\describe{
      \item{group}{recording the group ID}
      \item{frame}{recording the frame ID}
      \item{x}{x-axis variable}
      \item{y}{y-axis variable}
}}
\usage{
anidata
}
\description{
A Fake Data
}
\keyword{datasets}
