% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AniPlot.R
\name{AniPlot}
\alias{AniPlot}
\title{Animation Plot}
\usage{
AniPlot(x, GroupVar = "group", FrameVar = "frame", XVar = "x",
  YVar = "y", play.int = 0.5, AniPlot.dir = tempdir(),
  plotcaption = "My Animation Plot", plotwidth = 7, plotheight = 7,
  imagename = "My_Animation_Image", plotmain = NULL)
}
\arguments{
\item{x}{data to plot. It should be in data frame.}

\item{GroupVar}{the group variable name.}

\item{FrameVar}{the frame variable name.}

\item{XVar}{the x-axis variable name.}

\item{YVar}{the y-axis variable name.}

\item{play.int}{time interval to play each frame}

\item{AniPlot.dir}{directory to save the animation plot in PDF file}

\item{plotcaption}{animation plot caption}

\item{plotwidth}{animation plot width}

\item{plotheight}{animation plot height}

\item{imagename}{animation image file name}

\item{plotmain}{animation plot title}
}
\value{
Animation plot
}
\description{
\code{AniPlot} returns animation plots
}
\examples{
\dontrun{
AniPlot(anidata)
}
}
