% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdata.R
\docType{data}
\name{cdata}
\alias{cdata}
\title{First dose HPV vaccine uptake in current trial}
\format{A data frame with 475 rows and 3 variables:
\describe{
      \item{y}{first dose of hpv vaccine uptake}
      \item{x}{1 for treatment arm and 0 for control arm}
      \item{cl}{Each integer stands for a cluster}
}
source \url{https://clinicaltrials.gov/ct2/show/NCT02558803}}
\usage{
cdata
}
\description{
A dataset containing the first dose HPV
vaccine uptake information of 475 subjects from 28 clusters
}
\keyword{datasets}
