% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmodelBUGS.R
\name{cmodelBUGS}
\alias{cmodelBUGS}
\title{Bayesian mixed model analysis of current trial data}
\usage{
cmodelBUGS(cData, cForm, file.dir, file.name, file.rm = F, OpenBUGS.dir,
  OpenBUGS.seed = 1, family, nchain, niter, nburnin, nthin)
}
\arguments{
\item{cData}{current trial data. It should be in a data frame.}

\item{cForm}{model for current trial data. It should be in a formula object.}

\item{file.dir}{directory to save model BUGS file and input and output files of OpenBUGS.}

\item{file.name}{the name of the model BUGS file.}

\item{file.rm}{whether to remove all files after Bayesian computations are done.}

\item{OpenBUGS.dir}{directory to install OpenBUGS package.}

\item{OpenBUGS.seed}{random number generator state of OpenBUGS.}

\item{family}{outcome distribution.}

\item{nchain}{number of MCMC chains.}

\item{niter}{length of MCMC chain.}

\item{nburnin}{burnin period of MCMC chain.}

\item{nthin}{thinning rate of MCMC chain.}
}
\value{
The posterior samples of treatment effect
}
\description{
\code{cmodelBUGS} returns posterior samples of treatment effect
}
\examples{
\dontrun{
  ### my current data;
my.cData <- data.frame(cl=rep(1:10,each=50),y=rbinom(500,1,0.5),x=c(rep(1,250),rep(2,250)))

my.csample <- cmodelBUGS(cData=my.cData,cForm=y ~ x + (1|cl),
file.dir=tempdir(),file.name='cmodelfile',file.rm=T,
OpenBUGS.dir="C:/Program Files (x86)/OpenBUGS/OpenBUGS323/OpenBUGS.exe",
family="binomial",nchain=2,niter=500,nburnin=50,nthin=5)
}
}
