% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmodelfile.R
\name{cmodelfile}
\alias{cmodelfile}
\title{Generate current model BUGS file}
\usage{
cmodelfile(ntrt, family, mycfile)
}
\arguments{
\item{ntrt}{number of arms.}

\item{family}{outcome distribution.}

\item{mycfile}{the current model BUGS file.}
}
\value{
current model BUGS file
}
\description{
\code{cmodelfile} returns current model BUGS file
}
\examples{
\dontrun{
cmodelfile(4,"binomial",paste(tempdir(),"\\\\","cmodelfile.txt",sep=""))
cmodelfile(4,"gaussian",paste(tempdir(),"\\\\","cmodelfile.txt",sep=""))
cmodelfile(4,"poisson",paste(tempdir(),"\\\\","cmodelfile.txt",sep=""))
file.show(paste(tempdir(),"\\\\","cmodelfile.txt",sep=""))
}
}
