% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdata.R
\docType{data}
\name{hdata}
\alias{hdata}
\title{First dose HPV vaccine uptake in historical trial}
\format{A data frame with 475 rows and 5 variables:
\describe{
      \item{y}{number of subjects taking the first dose of HPV vaccine at the end of this trial}
      \item{n}{total number of subjects}
      \item{x}{arm status}
}}
\usage{
hdata
}
\description{
A dataset containing the first dose HPV
vaccine uptake information of 2139 patients in two arms
}
\keyword{datasets}
