% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmodelBUGS.R
\name{hmodelBUGS}
\alias{hmodelBUGS}
\title{Bayesian general linear regression analysis of current trial data}
\usage{
hmodelBUGS(hData, hForm, file.dir, file.name, file.rm = F, OpenBUGS.dir,
  OpenBUGS.seed = 1, family, nchain, niter, nburnin, nthin)
}
\arguments{
\item{hData}{historical trial data. It should be in a data frame.}

\item{hForm}{model for historical trial data. It should be a formula object.}

\item{file.dir}{directory to save model BUGS file and input and output files of OpenBUGS.}

\item{file.name}{the name of the model BUGS file.}

\item{file.rm}{whether to remove all files after Bayesian computations are done.}

\item{OpenBUGS.dir}{directory to install OpenBUGS package.}

\item{OpenBUGS.seed}{the random number generator state of OpenBUGS.}

\item{family}{outcome distribution.}

\item{nchain}{number of MCMC chains.}

\item{niter}{length of MCMC chain.}

\item{nburnin}{burnin period of MCMC chain.}

\item{nthin}{thinning rate of MCMC chain.}
}
\value{
The posterior samples of treatment effect
}
\description{
\code{hmodelBUGS} returns posterior samples of treatment effect
}
\examples{
\dontrun{
#' my historical data;
my.hData <- data.frame(y=rbinom(3,50,0.5),n=rep(50,3),x=c(1,2,3))

my.hsample <- hmodelBUGS(hData=my.hData,hForm=y|n ~ x,
file.dir=tempdir(),file.name='hmodelfile',file.rm=T, OpenBUGS.dir=NULL,
OpenBUGS.seed=14, family="binomial",nchain=2,niter=5000,nburnin=2500,nthin=5)
}
}
