% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmodelfile.R
\name{hmodelfile}
\alias{hmodelfile}
\title{Generate historical model BUGS file}
\usage{
hmodelfile(ntrt, family, myhfile)
}
\arguments{
\item{ntrt}{number of arms.}

\item{family}{outcome distribution.}

\item{myhfile}{the historical model BUGS file.}
}
\value{
historical model BUGS file
}
\description{
\code{hmodelfile} returns historical model BUGS file
}
\examples{
\dontrun{
hmodelfile(2,"binomial",paste(tempdir(),"\\\\","hmodelfile.txt",sep=""))
hmodelfile(3,"gaussian",paste(tempdir(),"\\\\","hmodelfile.txt",sep=""))
hmodelfile(4,"poisson",paste(tempdir(),"\\\\","hmodelfile.txt",sep=""))
file.show(paste(tempdir(),"\\\\","hmodelfile.txt",sep=""))
}
}
