% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprincrt.R
\docType{package}
\name{pprincrt}
\alias{pprincrt}
\alias{pprincrt-package}
\title{pprincrt: A Package for Design and Analysis of Cluster
          Randomized Trials}
\description{
The pprincrt package provides three types of important
functions: pprmodelBUGS, SimPower, AniPlot.
}
\section{pprmodelBUGS function}{

Analyze cluster randomized trial by borrowing information from
historical trial with power prior method, where the historical data
are summary data at the treatment group level
}

\section{SimPower function}{

Design cluster randomized trial under a historical parameter setting
or with a real historical data by using power prior method, where the
historical data should be summary one at the treatment group level. It
can also calculate the power of power prior method under a parameter
setting of both trials.
}

\section{AniPlot function}{

Create animation plot to inverstigate the relationship between four
variables.
}

