% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprmodelfile.R
\name{pprmodelfile}
\alias{pprmodelfile}
\title{Generate power prior method BUGS file}
\usage{
pprmodelfile(ntrt, family, mypprfile)
}
\arguments{
\item{ntrt}{number of arms.}

\item{family}{outcome distribution.}

\item{mypprfile}{the power prior method BUGS file.}
}
\value{
power prior method BUGS file
}
\description{
\code{pprmodelfile} returns power prior method BUGS file
}
\examples{
\dontrun{
pprmodelfile(4,"gaussian",paste(tempdir(),"\\\\","pprmodelfile.txt",sep=""))
pprmodelfile(3,"poisson",paste(tempdir(),"\\\\","pprmodelfile.txt",sep=""))
pprmodelfile(3,"binomial",paste(tempdir(),"\\\\","pprmodelfile.txt",sep=""))
file.show(paste(tempdir(),"\\\\","pprmodelfile.txt",sep=""))
}
}
