% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printpprMod.R
\name{print.pprMod}
\alias{print.pprMod}
\title{print pprMod object, returned by pprmodelBUGS function}
\usage{
\method{print}{pprMod}(x, open.plot = TRUE, ...)
}
\arguments{
\item{x}{pprMod object. A S3 object defined by this package developer.}

\item{open.plot}{An indicator of opening the convergence diagnostic plots. It is TRUE by default.}

\item{\dots}{other arguments}
}
\value{
summary and convergence diagnositc results of MCMC chains
}
\description{
\code{pprmodelBUGS} returns summary and convergence diagnositc results of MCMC chains
}
