% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPower.R
\name{SimPower}
\alias{SimPower}
\title{power in designing a new cluster randomized trial with the power prior method}
\usage{
SimPower(family, to.do.option, cSet, cNarmVar = "narm",
  cNcluster.armVar = "ncluster.arm", cNpat.clusterVar = "npat.cluster",
  cBtwVar = "sigma.b", cWthVar = "sigma.e", cTrtVecVar, hSet = NULL,
  hNarmVar = "narm", hNpat.armVar = "npat.arm", hWthVar = "sigma.e",
  hTrtVecVar = NULL, hData = NULL, hTrtVar = "x", hOutcomeVar = "y",
  hSumVar = "n", cRdmSeed.init = 0, hRdmSeed.init = 0, weight = "asym",
  cover.level = 0.95, nchain = 1, niter = 2000, nburnin = 1000,
  nthin = 1, file.dir.init = tempdir(), cfile.name = "cmodelfile",
  hfile.name = "hmodelfile", pprfile.name = "pprmodelfile",
  graphics.name = "graphics", file.rm = T, OpenBUGS.dir = NULL,
  OpenBUGS.seed = 1, Rep = 1, num.core = parallel::detectCores(logical =
  F) - 1)
}
\arguments{
\item{family}{outcome distribution.}

\item{to.do.option}{a value in 'sim_power', 'sim_SSD' and 'real_SSD'.}

\item{cSet}{current trial setting. It should be a list.}

\item{cNarmVar}{the variable name of arm in current trial.}

\item{cNcluster.armVar}{the variable name of total clusters in each arm in current trial. Its default value is 'ncluster.arm'.}

\item{cNpat.clusterVar}{the variable name of total patients in each cluster in current trial. Its default value is 'npat.cluster'.}

\item{cBtwVar}{the variable name of between cluster variation in current trial.}

\item{cWthVar}{the variable name of within cluster variation in current trial.}

\item{cTrtVecVar}{the variable name of treatment in current trial. It should be in the order of first treatment, second treatment until the last one.}

\item{hSet}{historical trial setting. It should be in a list. It is NULL by default.}

\item{hNarmVar}{the variable name of arm in historical trial.}

\item{hNpat.armVar}{the variable name of total patients in each arm in historical trial. Its default value is 'npat.arm'.}

\item{hWthVar}{the variable name of within cluster variation in historical trial.}

\item{hTrtVecVar}{the variable name of treatment in historical trial. It should be in the order of first treatment, second treatment until the last one.}

\item{hData}{Real historical trial data. It is NULL by default.}

\item{hTrtVar}{the variable name of treatment in historical trial data.}

\item{hOutcomeVar}{the variable name of outcome in historical trial data.}

\item{hSumVar}{the variable name of total observation in historical trial data.}

\item{cRdmSeed.init}{the initial value of random seed to generate current trial data.}

\item{hRdmSeed.init}{the initial value of random seed to generate historical trial data.}

\item{weight}{discounting parameter. It is 'asym' by default.}

\item{cover.level}{nomial coverage level of HPD interval.}

\item{nchain}{number of mcmc chains to run.}

\item{niter}{length of mcmc chains.}

\item{nburnin}{length of burnin of mcmc chains.}

\item{nthin}{thinning rate of mcmc chain.}

\item{file.dir.init}{the start of the directory to save the model BUGS file and input and output files of OpenBUGS.}

\item{cfile.name}{the file name of current BUGS model.}

\item{hfile.name}{the file name of historical BUGS model.}

\item{pprfile.name}{the file name of power prior BUGS model.}

\item{graphics.name}{the file name of mcmc convergence graphics.}

\item{file.rm}{whether to remove all files after Bayesian computations are done.}

\item{OpenBUGS.dir}{directory to install OpenBUGS package.}

\item{OpenBUGS.seed}{the random number generator state of OpenBUGS.}

\item{Rep}{number of replications to determine power.}

\item{num.core}{number of cores for parallel computation.}
}
\value{
power in designing a new cluster randomized trial with the power prior method
}
\description{
\code{SimPower} returns power in designing a new cluster randomized trial with the power prior method
}
\examples{
\dontrun{
#### SSD with simulated historical data;
my.curset <- list(narm=2,ncluster.arm=rep(10,2),npat.cluster=rep(20,20),
'trt1'=-0.1,'trt2 v.s. trt1'=log(4),sigma.b=1)
my.histset <- list(narm=2,npat.arm=rep(200,2),'trt1'=0.1,'trt2 v.s. trt1'=log(4))
mypower <- SimPower(family='binomial', to.do.option='sim_SSD', cSet=my.curset,
cTrtVecVar=c('trt1','trt2 v.s. trt1'), hSet=my.histset,
hTrtVecVar=c('trt1','trt2 v.s. trt1'),
weight="asym",nchain=2,niter=500,nburnin=100,nthin=5,Rep=1,num.core=3,file.rm=T,
OpenBUGS.dir="C:\\\\Program Files (x86)\\\\OpenBUGS\\\\OpenBUGS323\\\\OpenBUGS.exe",
OpenBUGS.seed=12)
 #### SSD with real historical data;
my.curset <- list(narm=2,ncluster.arm=rep(10,2),npat.cluster=rep(20,20),
'trt1'=-0.1,'trt2 v.s. trt1'=log(4),sigma.b=1)
my.histdata <- data.frame(x=c(1,2),y=c(10,15),n=c(20,20))
mypower <- SimPower(family='binomial', to.do.option='real_SSD',cSet=my.curset,
cTrtVecVar=c('trt1','trt2 v.s. trt1'),hData=my.histdata,weight='sym',
nchain=2,niter=500,nburnin=100,nthin=5,file.rm=T, Rep=4,num.core=2,
OpenBUGS.dir="C:\\\\Program Files (x86)\\\\OpenBUGS\\\\OpenBUGS323\\\\OpenBUGS.exe")
}
}
