% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimRej.R
\name{SimRej}
\alias{SimRej}
\title{rejection decision of null treatment effect}
\usage{
SimRej(family, to.do.option, cSet, cNarmVar = "narm",
  cNcluster.armVar = "ncluster.arm", cNpat.clusterVar = "npat.cluster",
  cTrtVecVar, cBtwVar = "sigma.b", cWthVar = "sigma.e", hSet = NULL,
  hNarmVar = "narm", hNpat.armVar = "npat.arm", hTrtVecVar = NULL,
  hWthVar = "sigma.e", hData = NULL, hTrtVar = "x", hOutcomeVar = "y",
  hSumVar = "n", cRdmSeed = 1, hRdmSeed = 1, weight = "asym",
  cover.level = 0.95, nchain = 1, niter = 2000, nburnin = 1000,
  nthin = 1, file.dir = tempdir(), cfile.name = "cmodelfile",
  hfile.name = "hmodelfile", pprfile.name = "pprmodelfile",
  graphics.name = "graphics", file.rm = F, OpenBUGS.dir = NULL,
  OpenBUGS.seed = 1)
}
\arguments{
\item{family}{outcome distribution}

\item{to.do.option}{a value in 'sim_power', 'sim_SSD' and 'real_SSD'}

\item{cSet}{current trial setting. It should be in a list.}

\item{cNarmVar}{the variable name of arm in current trial.}

\item{cNcluster.armVar}{the variable name of total clusters in each arm in current trial. Its default value is 'ncluster.arm'.}

\item{cNpat.clusterVar}{the variable name of total patients in each cluster in current trial. Its default value is 'npat.cluster'.}

\item{cTrtVecVar}{the variable name of treatment in current trial. It should be in the order of first treatment, second treatment until the last one.}

\item{cBtwVar}{the variable name of between cluster variation in current trial.}

\item{cWthVar}{the variable name of within cluster variation in current trial.}

\item{hSet}{historical trial setting. It should be in a list. It is NULL by default.}

\item{hNarmVar}{the variable name of arm in historical trial.}

\item{hNpat.armVar}{the variable name of total patients in each arm in historical trial. Its default value is 'npat.arm'.}

\item{hTrtVecVar}{the variable name of treatment in historical trial. It should be in the order of first treatment, second treatment until the last one.}

\item{hWthVar}{the variable name of within cluster variation in historical trial.}

\item{hData}{Real historical trial data. It is NULL by default.}

\item{hTrtVar}{the variable name of treatment in historical trial data.}

\item{hOutcomeVar}{the variable name of outcome in historical trial data.}

\item{hSumVar}{the variable name of total observation in historical trial data.}

\item{cRdmSeed}{the random seed to generate current trial data.}

\item{hRdmSeed}{the random seed to generate historical trial data.}

\item{weight}{discounting parameter. It is 'asym' by default.}

\item{cover.level}{nomial coverage level of HPD interval.}

\item{nchain}{number of mcmc chains to run.}

\item{niter}{length of mcmc chains.}

\item{nburnin}{length of burnin of mcmc chains.}

\item{nthin}{thinning rate of mcmc chain.}

\item{file.dir}{directory to save the model BUGS file and input and output files of OpenBUGS.}

\item{cfile.name}{the file name of current BUGS model.}

\item{hfile.name}{the file name of historical BUGS model.}

\item{pprfile.name}{the file name of power prior BUGS model.}

\item{graphics.name}{the file name of mcmc convergence graphics.}

\item{file.rm}{whether to remove all files after Bayesian computations are done.}

\item{OpenBUGS.dir}{directory to install OpenBUGS package.}

\item{OpenBUGS.seed}{the random number generator state of OpenBUGS.}
}
\value{
rejection decision of null treatment effect
}
\description{
\code{SimRej} returns rejection decision of null treatment effect
}
\examples{
\dontrun{
my.curset <- list(narm=3,ncluster.arm=rep(5,3),npat.cluster=rep(20,15),
'trt1'=0.1,'trt2 v.s. trt1'=1,'trt3 v.s. trt1'=1,sigma.b=1)
my.histset <- list(narm=3,npat.arm=rep(20,3),
'trt1'=-0.1,'trt2 v.s. trt1'=1,'trt3 v.s. trt1'=1)
myrej <- SimRej(family='binomial', to.do.option="sim_power",cSet=my.curset,
cTrtVecVar=c('trt1', paste(paste('trt', 2:3, sep=""),'v.s.', 'trt1',sep=" ")),
hSet=my.histset,
hTrtVecVar=c('trt1', paste(paste('trt', 2:3, sep=""),'v.s.', 'trt1',sep=" ")),
file.rm=T, nchain=2,niter=200,nburnin=50,nthin=5, OpenBUGS.seed=4)
}
}
